+++ base/exports.jl
@@ -1171,8 +1171,6 @@ export
 
     # output
     @show,
-    @printf,
-    @sprintf,
 
     # profiling
     @time,