+++ base/printf.jl
@@ -3,7 +3,6 @@
 module Printf
 using Base: Grisu, GMP
 using Base.Unicode: lowercase, textwidth, isupper
-export @printf, @sprintf
 
 ### printf formatter generation ###
 const SmallFloatingPoint = Union{Float64,Float32,Float16}
@@ -1193,27 +1192,6 @@ function _printf(macroname, io, fmt, args)
     Expr(:let, Expr(:block), blk)
 end
 
-"""
-    @printf([io::IOStream], "%Fmt", args...)
-
-Print `args` using C `printf` style format specification string, with some caveats:
-`Inf` and `NaN` are printed consistently as `Inf` and `NaN` for flags `%a`, `%A`,
-`%e`, `%E`, `%f`, `%F`, `%g`, and `%G`. Furthermore, if a floating point number is
-equally close to the numeric values of two possible output strings, the output
-string further away from zero is chosen.
-
-Optionally, an `IOStream`
-may be passed as the first argument to redirect output.
-
-# Examples
-```jldoctest
-julia> @printf("%f %F %f %F\\n", Inf, Inf, NaN, NaN)
-Inf Inf NaN NaN\n
-
-julia> @printf "%.0f %.1f %f\\n" 0.5 0.025 -0.0078125
-1 0.0 -0.007813
-```
-"""
 macro printf(args...)
     isempty(args) && throw(ArgumentError("@printf: called with no arguments"))
     if isa(args[1], AbstractString) || is_str_expr(args[1])
@@ -1225,19 +1203,6 @@ macro printf(args...)
     end
 end
 
-"""
-    @sprintf("%Fmt", args...)
-
-Return `@printf` formatted output as string.
-
-# Examples
-```jldoctest
-julia> s = @sprintf "this is a %s %15.1f" "test" 34.567;
-
-julia> println(s)
-this is a test            34.6
-```
-"""
 macro sprintf(args...)
     isempty(args) && throw(ArgumentError("@sprintf: called with zero arguments"))
     isa(args[1], AbstractString) || is_str_expr(args[1]) ||